adicionado helpers

+++ library/src/main/java/bm/it/mobile/ui/helpers/ViewHelpers.kt
@@ -5,7 +5,7 @@ import android.util.DisplayMetrics
 
 object ViewHelpers {
 
-    fun calculateNoOfColumns(context: Context, columnSize: Int): Int {
+    fun calculateNoOfColumns(context: Context, columnSize: Float): Int {
         val displayMetrics: DisplayMetrics = context.resources.displayMetrics
         val screenWidthDp = displayMetrics.widthPixels / displayMetrics.density
         return (screenWidthDp / columnSize + 0.5).toInt()
